[skip si] moved appitem state and description to app item folder
