+++ DearPyGui/src/core/PythonCommands/mvAppInterface.cpp
@@ -1,6 +1,6 @@
 #include "mvAppInterface.h"
 #include "mvInterfaceCore.h"
-#include "core/mvAppItems.h"
+#include "core/AppItems/mvAppItems.h"
 #include "core/AppItems/mvWindowAppItem.h"
 #include "core/mvWindow.h"
 #include "core/mvEvents.h"
