+++ DearPyGui/src/core/PythonCommands/mvSliderInterface.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "core/mvAppItems.h"
+#include "core/AppItems/mvAppItems.h"
 
 namespace Marvel {
